+++ src-tauri/mc-laucher-lib-rs/src/lib.rs
@@ -13,6 +13,10 @@ mod command;
 pub mod login;
 pub mod client;
 
+pub use install::swap_mods_folder;
+pub use install::install_mods;
+pub use install::update_mods;
+pub use install::Mod;