+++ src/lib/db.ts
@@ -1,4 +1,5 @@
 import { Db } from 'zangodb';
+import { RemoveModsFolder } from './invoke';
 import { Profile } from '../types';
 export default class DB {
     static INSTANCE: DB | null = null;
@@ -15,10 +16,7 @@ export default class DB {
     constructor(){
         if(DB.INSTANCE) return DB.INSTANCE;
         this.db = new Db("minecraft",1,{ profiles: ["uuid"], mods: ["uuid"], users: ["xuid"] });
-        DB.INSTANCE = this;
-        //@ts-ignore
-        window.db = this;
-        
+        DB.INSTANCE = this; 
     }
     get users(){
         return this.db.collection("users");
@@ -28,5 +26,30 @@ export default class DB {
     }
     get mods(){
         return this.db.collection("mods");
+    }
+    public async Clear(){
+        this.db.drop();
+
+        const profiles = await this.profiles.find({}).toArray() as Profile[];
+
+        for(const pro of profiles){
+            this.dropProfile(pro.uuid);
+        }
+     
+        DB.Get();
+    }
+    public async removeMod(profile: string, uuid: string): Promise<void> {
+        
+        // native
+        const data = await this.profiles.findOne({ uuid: profile }) as Profile;
+
+        const mods = data.mods.filter(i=>i.id !== uuid);
+
+        await this.profiles.update({ uuid: profile }, { mods });
+    }
+    public async dropProfile(uuid: string): Promise<void> {
+
+        await RemoveModsFolder(uuid);
+        await this.profiles.remove({ uuid });
     }
 }